fix: get full animal's age
