+++ packages/design-system/helpers/getAge/index.js
@@ -1,14 +1,37 @@
-const getAge = (date) => {
-  const birthDate = new Date(date)
-  const year = birthDate.getFullYear()
-  const month = birthDate.getMonth()
-  const day = birthDate.getDay()
+import getAgeNumber from "../getAgeNumber"
 
-  const birth = new Date(year, month - 1, day)
-  const now = new Date()
-  const diff = new Date(now.valueOf() - birth.valueOf())
+export function getPlurals(
+  singularNominativ,
+  pluralNominativ,
+  pluralGenitive,
+  value
+) {
+  const newVal = Math.abs(value)
 
-  return Math.abs(diff.getFullYear() - 1970)
+  if (newVal === 1) {
+    return singularNominativ
+  }
+  if (
+    newVal % 10 >= 2 &&
+    newVal % 10 <= 4 &&
+    (newVal % 100 < 10 || newVal % 100 >= 20)
+  ) {
+    return pluralNominativ
+  }
+  return pluralGenitive
+}
+
+const getAge = (age) => {
+  if (getAgeNumber(age) < 1) {
+    return `poniżej roku`
+  }
+
+  return `~ ${getAgeNumber(age)} ${getPlurals(
+    "rok",
+    "lata",
+    "lat",
+    getAgeNumber(age)
+  )}`
 }
 
 export default getAge
