+++ packages/design-system/helpers/getAgeNumber/__tests__/getAgeNumber.test.js
@@ -1,25 +1,25 @@
-import getAge from "design-system/helpers/getAge"
+import getAgeNumber from "design-system/helpers/getAgeNumber"
 
 const testCases = [
   {
     value: "2021-03-24T00:00:00.000Z",
-    expected: "~ 2 lata",
+    expected: 2,
   },
   {
     value: "2022-07-24T00:00:00.000Z",
-    expected: "poniżej roku",
+    expected: 0,
   },
   {
     value: "2015-03-24T00:00:00.000Z",
-    expected: "~ 8 lat",
+    expected: 8,
   },
 ]
 
-describe(`getAge`, () => {
+describe(`getAgeNumber`, () => {
   test.each(testCases)(
     "returns $expected from the date: $value",
     ({ value, expected }) => {
-      expect(getAge(value)).toBe(expected)
+      expect(getAgeNumber(value)).toBe(expected)
     }
   )
 })
