+++ packages/frontend/app/templates/Animal/Animal.jsx
@@ -35,7 +35,7 @@ const Animal = ({ data }) => (
       <Container>
         <Details
           gender={data.gender}
-          age={data.age2}
+          age={data.age}
           height={data.height}
           weight={data.weight}
           targetWeight={data.targetWeight}
